fix: widen the HZ5.2 image framebuffer offset to 32 bits
DirectPixelWriter computed its framebuffer byte offset as uint16_t. HZ5.2's
framebuffer is 1280x720/8 = 115200 B, so every offset above 65535 wrapped and
the affected pixels landed 409 physical rows up and 96 bytes across -- the wrap
point falls mid-row, so the displacement is in both axes. Full-page images drew
as two overlapping copies at different origins.

Measured on device with temporary instrumentation: one full-page image reported
writes=30358 wrapped=7898 maxIndex=95447, a quarter of its pixels truncated.

X4 (48000 B) and M4 (58080 B) both fit in 16 bits, so this is the first panel to
expose it. Text was never affected because GfxRenderer::drawPixel computes the
identical offset in uint32_t; the two copies of this write path diverged on
integer width, not on geometry. The orientation transform is not implicated here
and still has exactly two copies, both correct.

The same run also cleared the other two suspects: stripOriginY=0 stripRows=720
(strip target inactive) and exactly one render entry per page (the BW pass runs
once).

+++ lib/Epub/Epub/converters/DirectPixelWriter.h
@@ -175,7 +175,11 @@ struct DirectPixelWriter {
     const int sy = phyY - originY;
     if (static_cast<unsigned>(sy) >= static_cast<unsigned>(clipRows)) return;
 
-    const uint16_t byteIndex = static_cast<uint16_t>(sy * displayWidthBytes + (phyX >> 3));
+    // 32-bit: a framebuffer offset does not fit in 16 bits on a large panel. HZ5.2 is
+    // 1280x720 = 115200 B, so every physical row past 409 wrapped back to the top and drew
+    // a second, displaced copy of the image. Matches GfxRenderer::drawPixel(), which has
+    // always computed this offset in 32 bits -- which is why text was unaffected.
+    const uint32_t byteIndex = static_cast<uint32_t>(sy) * displayWidthBytes + (phyX >> 3);
     const uint8_t bitMask = 1 << (7 - (phyX & 7));
 
     if (state) {
